requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
ffmpeg-python==0.2.0
frozenlist==1.4.0
future==0.18.3
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
numpy==1.25.2
openai==0.27.8
pycparser==2.21
pydantic==1.10.12
python-dotenv==0.21.1
pytz==2023.3
requests==2.31.0
SpeechRecognition==3.10.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
